buffacli/pyproject.toml
[project]
name = "buffacli"
version = "0.0.1"
description = "A CLI tool for interacting with BuffaLogs"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer==0.16.0",
    "requests==2.32.3",
    "rich==14.0.0",
    "yarl==1.20.1",
    "typing-extensions==4.14.0",
    "pydantic>=2.11.10",
    "python-dateutil>=2.9.0.post0",
    "pytest>=8.4.2",
]

[project.scripts]
buffacli = "buffacli.buffacli:app"  # Exposes `buffacli` command via Typer

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
